[hotfix] [streaming] Move SerializedCheckpointData to proper scope for MessageAcknowledgingSourceBase
